+++ src/App.tsx
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import {BrowserRouter as Router, Route, Link } from 'react-router-dom';
-import logo from './logo.svg';
 import './App.css';
+import HomePage from './pages/HomePage';
+import LoginPage from './pages/LoginPage';
 
-class AppRouter extends Component {
+class App extends Component {
   render() {
     return (
       <Router>
@@ -16,46 +17,4 @@ class AppRouter extends Component {
   }
 }
 
-interface User {
-  email: string;
-}
-
-class HomePage extends Component {
-  readonly state: {user: User | null} = {user: null};
-
-  componentDidMount() {
-    this.fetchUser();
-  }
-
-  fetchUser() {
-    fetch('/current-user')
-      .then((response) => response.json())
-      .then((data: {user: User}) => {
-        this.setState({user: data.user});
-      });
-  }
-
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1>Kvasir Movies</h1>
-          <p>
-            {this.state.user != null ? `Hi ${this.state.user.email}! ` : ''}
-            Find 🎬 with 👫 :D
-          </p>
-          <a href='/login'>Log In</a>
-        </header>
-      </div>
-    );
-  }
-}
-
-class LoginPage extends Component {
-  render() {
-    return <div>Log in stuff will go here</div>
-  }
-}
-
-export default AppRouter;
+export default App;
