+++ src/pages/LoginPage.tsx
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class LoginPage extends Component {
+    render() {
+      return <div>Log in stuff will go here</div>
+    }
+  }
+
+  export default LoginPage;
